refactor: split Sentry observability helpers
Move Sentry redaction constants and pure helper logic out of the Sentry entrypoint so configuration and capture wiring stay focused.
